backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
pandas==2.1.3
duckdb==0.9.2
pymilvus==2.3.3
sentence-transformers==2.2.2
